+++ packages/config/__tests__/Config.api.test.ts
@@ -11,9 +11,12 @@
 
 import * as fs from "fs";
 import * as path from "path";
+import * as JSONC from "comment-json";
 import { ConfigSecure } from "../src/api";
 import { Config } from "../src/Config";
+import { ConfigConstants } from "../src/ConfigConstants";
 import { IConfig } from "../src/doc/IConfig";
+import { IConfigLayer } from "../src/doc/IConfigLayer";
 import { IConfigProfile } from "../src/doc/IConfigProfile";
 
 const MY_APP = "my_app";
@@ -423,6 +426,29 @@ describe("Config API tests", () => {
                 expect(retrievedConfig.profiles.fruit.profiles.apple.properties.color).toBe("red");
                 expect(retrievedConfig.profiles.fruit.profiles.orange).toBeDefined();
                 expect(retrievedConfig.profiles.fruit.properties.origin).toBe("California");
+            });
+        });
+        describe("merge - dry run", () => {
+            it("should merge config layers with correct priority", async () => {
+                const config = await Config.load(MY_APP);
+                const existingConfig = JSONC.parse(JSONC.stringify(config.layerActive(), null, ConfigConstants.INDENT));
+                const retrievedConfig = (config.api.layers.merge(mergeConfig, true) as IConfigLayer).properties;
+                expect(retrievedConfig).toMatchSnapshot();
+
+                // Check that new config was added
+                expect(retrievedConfig.plugins.length).toBe(2);
+                expect(retrievedConfig.profiles.fruit.profiles.grape).toBeDefined();
+                expect(retrievedConfig.profiles.fruit.properties.shipDate).toBeDefined();
+                expect(retrievedConfig.profiles.fruit.secure.length).toBe(1);
+
+                // Check that old config had priority
+                expect(retrievedConfig.defaults.fruit).toBe("fruit.apple");
+                expect(retrievedConfig.profiles.fruit.profiles.apple.properties.color).toBe("red");
+                expect(retrievedConfig.profiles.fruit.profiles.orange).toBeDefined();
+                expect(retrievedConfig.profiles.fruit.properties.origin).toBe("California");
+
+                // Check that the original was not modified
+                expect(config.layerActive()).toEqual(existingConfig);
             });
         });
     });