+++ packages/imperative/src/doc/IImperativeConfig.ts
@@ -14,6 +14,7 @@ import { IImperativeLogsConfig } from "./IImperativeLogsConfig";
 import { IImperativeOverrides } from "./IImperativeOverrides";
 import { IImperativeAuthGroupConfig } from "./IImperativeAuthGroupConfig";
 import { IApimlSvcAttrs } from "./IApimlSvcAttrs";
+import { ICommandProfileAutoInitConfig } from "../../../cmd/src/doc/profiles/definition/ICommandProfileAutoInitConfig";
 
 /**
  * All of the configuration required to set up your Imperative CLI app
@@ -174,6 +175,13 @@ export interface IImperativeConfig {
      * @memberof IImperativeConfig
      */
     authGroupConfig?: IImperativeAuthGroupConfig;
+
+    /**
+     * Use this property to customize the command definition for the config init command.
+     * @type {ICommandProfileAutoInitConfig}
+     * @memberof IImperativeConfig
+     */
+    configAutoInitCommandConfig?: ICommandProfileAutoInitConfig;
 
     /**
      * Specify the name to use in the config template for the root profile that contains all supported profiles.